添加RecyclerView分割线装饰器

+++ baseDevKit/src/main/java/com/dev/kit/basemodule/util/DisplayUtil.java
@@ -101,7 +101,7 @@ public class DisplayUtil {
         return mDisplayUtil.mFontScaleFactor;
     }
 
-    public static int dp2px(int dpValue) {
+    public static int dp2px(float dpValue) {
         if (mDisplayUtil == null) {
             getInstance();
         }
